fixed bugs quizzdata.score
