+++ lib/services/auth/facebook_auth.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 import 'package:flutter_facebook_login/flutter_facebook_login.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert' as JSON;
@@ -15,7 +14,6 @@ bool isFacebookLogined = false;
 
 Future<Null> signInWithFacebook() async {
   final FacebookLoginResult result = await facebookSignIn.logIn(['email']);
-
   switch (result.status) {
     case FacebookLoginStatus.loggedIn:
       final accessToken = result.accessToken.token;
