+++ lib/widgets/answer_list.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:history_quizz/models/quizz_brain.dart';
-import 'package:history_quizz/widgets/question_cards.dart';
-import 'package:provider/provider.dart';
-import 'answer_cards.dart';
+import 'package:history_quizz/views/answer_cards.dart';
 
 class AnswerList extends StatelessWidget {
   AnswerList(
@@ -21,17 +18,8 @@ class AnswerList extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<QuizzData>(
-      builder: (context, quizzdata, child) {
-        return Column(
-          children: <Widget>[
-            answerText1,
-            answerText2,
-            answerText3,
-            answerText4
-          ],
-        );
-      },
+    return Column(
+      children: <Widget>[answerText1, answerText2, answerText3, answerText4],
     );
   }
 }
