+++ lib/widgets/drawer_list_titles.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:history_quizz/models/quizz_brain.dart';
-import 'package:history_quizz/screens/main_screen.dart';
+import 'package:history_quizz/models/providers/quizz_brain.dart';
 import 'package:provider/provider.dart';
 
 class DrawerListTitle extends StatelessWidget {
