+++ lib/widgets/header.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:history_quizz/models/quizz_brain.dart';
+import 'package:history_quizz/models/providers/quizz_brain.dart';
 import 'package:history_quizz/models/user_score.dart';
-import 'package:history_quizz/screens/main_screen.dart';
-import 'package:history_quizz/screens/welcome_screen.dart';
-import 'package:history_quizz/utils/score_data.dart';
+import 'package:history_quizz/models/utils/score_data.dart';
+import 'package:history_quizz/pages/welcome_screen.dart';
 import 'package:provider/provider.dart';
 
 class Header extends StatefulWidget {
@@ -21,6 +20,7 @@ class _HeaderState extends State<Header> {
     newScore(
       Score(score: '${Provider.of<QuizzData>(context, listen: false).score}'),
     );
+    print('${Provider.of<QuizzData>(context, listen: false).score}');
     super.initState();
   }
 
@@ -42,43 +42,48 @@ class _HeaderState extends State<Header> {
                         ),
                         onTap: () {
                           quizzdata.score -= 3;
-                    newScore(
-                      Score(
-                          score:
-                          '${Provider
-                              .of<QuizzData>(context, listen: false)
-                              .score}'),
-                    );
-                    print(getScore());
-                    Navigator.of(context).popAndPushNamed(WelcomePage.id);
-                  },
-                ),
-                Text(
-                  widget.time,
-                  style: TextStyle(
-                    fontSize: 32.0,
-                    color: Color(0xFFFFFFFF),
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-                Stack(
-                  alignment: Alignment.center,
-                  children: <Widget>[
-                    Icon(
-                      Icons.favorite_border,
-                      size: 62.0,
-                    ),
-                    Text(
-                      snapshot.data.last.score,
-                      style: TextStyle(
-                        fontSize: 20.0,
-                        color: Color(0xFFFFFFFF),
-                        fontWeight: FontWeight.bold,
+                          newScore(
+                            Score(
+                                score:
+                                '${Provider
+                                    .of<QuizzData>(context, listen: false)
+                                    .score}'),
+                          );
+                          print(getScore());
+                          Navigator.of(context).popAndPushNamed(WelcomePage.id);
+                        },
                       ),
-                    ),
-                  ],
-                )
-              ],
+                      Text(
+                        widget.time,
+                        style: TextStyle(
+                          fontSize: 32.0,
+                          color: Color(0xFFFFFFFF),
+                          fontWeight: FontWeight.w500,
+                        ),
+                      ),
+                      GestureDetector(
+                        onTap: () {
+                          print('${snapshot.data.last.score}');
+                        },
+                        child: Stack(
+                          alignment: Alignment.center,
+                          children: <Widget>[
+                            Icon(
+                              Icons.favorite_border,
+                              size: 62.0,
+                            ),
+                            Text(
+                              quizzdata.score.toString(),
+                              style: TextStyle(
+                                fontSize: 20.0,
+                                color: Color(0xFFFFFFFF),
+                                fontWeight: FontWeight.bold,
+                              ),
+                            ),
+                          ],
+                        ),
+                      )
+                    ],
             );
           },
         )
